refactor: UITextView Extension에서 기존의 중복 코드를 처리하도록 변경
